// src/types/stock.ts
export type StockOperationEntity = {
  id?: number;
  bankAccountId: number;
  bankAccountDescription?: string;
  symbol: string;
  operationDate: string;
  operationType: StockOperationType;
  quantity: number;
  price: number;
  total: number;
};

export enum StockOperationType {
  Purchase = "PURCHASE",
  Sell = "SELL",
}

// Generated by https://quicktype.io

export interface StockWatchListEntity {
  id?: number;
  thumbImage?: string;
  symbol: string;
  description: string;
  active: boolean;
  lowerTarget: number;
  upperTarget: number;
  currentPrice: number;
}